refactor statuspagecomponents into own module
split into files representing models from qontract-schema and the
atlassian specific things
