+++ reconcile/test/fixtures/statuspage/test_bind_component.yaml
@@ -20,7 +20,7 @@ atlassianApi:
     page_1:
     - id: comp_id_1
       name: Component 1
-      descripton: null
+      description: null
       position: 1
       status: Ok
       automation_email: automate-me@please.com
@@ -29,7 +29,7 @@ atlassianApi:
       group_name: group_1
     - id: group_id_1
       name:  group_1
-      descripton: null
+      description: null
       position: 1
       status: Ok
       automation_email: null
