+++ reconcile/test/fixtures/statuspage/test_component_status_update.yaml
@@ -22,7 +22,7 @@ atlassianApi:
     page_1:
     - id: comp_id_1
       name: Component 1
-      descripton: null
+      description: null
       position: 1
       status: Ok
       automation_email: automate-me@please.com
